% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/objectives.R
\name{lsq_obj}
\alias{lsq_obj}
\title{Least-squares objective function}
\usage{
lsq_obj(a, b, t)
}
\arguments{
\item{a}{the n * p matrix A}

\item{b}{the n * q matrix B}

\item{t}{the nuclear norm constrant on X, doubled}
}
\value{
A list with elements \code{f}, the function, \code{gradf}, the gradient,
\code{cf}, the curvature constant,
\code{convert}, a function which converts the solution of the transformed problem to the original,
\code{original_f}, computes the original objective function and nuclear norm
}
\description{
Furnishes closures which evaluate the function and gradient for
the transformation of the problem min_X tr((AX - B)'(AX - B))
subject to NuclearNorm(X) <= t/2
}

