% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuclear_rate3Hedp_5p.R
\name{nuclear_rate3Hedp_5p}
\alias{nuclear_rate3Hedp_5p}
\title{Estimate reaction rates}
\format{\describe{
\item{x}{
The function has four arguments: ECM, gd, gp, rd, rp, T9}
}}
\usage{
nuclear_rate3Hedp_5p(ECM = ECM, gd = gd, gp = gp, rd = rd, rp = rp, T9 = T9)
}
\arguments{
\item{e0}{E0}

\item{er}{ER}

\item{gi}{gi}

\item{gf}{gf}

\item{ri}{ri}

\item{rf}{rf}

\item{T9}{T9}

\item{rp}{rp}
}
\value{
nuclear_rate
}
\description{
Calculate numerical reaction rates
}
\examples{
library(nuclear)

N <- 300
obsx1 <- exp(seq(log(1e-3), log(1),length.out=N))
plot(obsx1,sfactor3Hedp(obsx1,0.35,1.0085,0.025425),
col="red",cex=1.25,type="l",ylab="S-factor",xlab="E",log="x")

}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{Nuclear_rate}
