% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRate3HedpLoop.R
\name{NumRate3HedpLoop}
\alias{NumRate3HedpLoop}
\title{Estimate reaction rates over a temperature grid}
\format{\describe{
\item{x}{
The function has 4 arguments: mat, vars, N, T9}
}}
\usage{
NumRate3HedpLoop (mat = mat, vars=vars, N = N, T9  =T9)
}
\arguments{
\item{mat}{mat}

\item{vars}{vars}

\item{N}{N}

\item{T9}{T9}
}
\value{
NumRate3HedpLoop
}
\description{
Provides a reaction rate table
}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{NumRate3HedpLoop}
