% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRate7BenpTable.R
\name{NumRate7BenpTable}
\alias{NumRate7BenpTable}
\alias{NumRateHe3dpTable}
\title{Estimate reaction rates over a temperature grid}
\format{\describe{
\item{x}{
The function has 4 arguments: mat, vars, N, T9}
}}
\usage{
NumRateHe3dpTable (mat = mat, vars=vars, N = N, T9  =T9)
}
\arguments{
\item{mat}{mat}

\item{N}{N}

\item{T9}{T9}

\item{vars}{vars}
}
\value{
NumRate7BenpTable
}
\description{
Provides a reaction rate table
}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{NumRate7BenpTable}
