% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRateDpg.R
\name{NumRateDpg}
\alias{NumRateDpg}
\title{Estimate reaction rates}
\format{\describe{
\item{x}{
The function has two  arguments: a.scale, T9}
}}
\usage{
NumRateDpg(a.scale = a.scale,T9 = T9)
}
\arguments{
\item{a.scale}{a.scale}

\item{T9}{T9}
}
\value{
nuclear_rate
}
\description{
Calculate numerical reaction rates
}
\examples{
library(nuclear)

 NumRateDpg(1,10)

}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{Nuclear}
\keyword{rate}
