% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRateHe3He3.R
\name{NumRateHe3He3}
\alias{NumRateHe3He3}
\title{Estimate reaction rates}
\format{\describe{
\item{x}{
The function has two  arguments: S0,S1,S2,i.screen, T9}
}}
\usage{
NumRateHe3He3(S0 = S0,S1 = S1,S2 = S2,i.screen = i.screen,T9 = T9)
}
\arguments{
\item{S0}{S0}

\item{S1}{S1}

\item{S2}{S2}

\item{i.screen}{i.screen}

\item{T9}{T9}
}
\value{
nuclear_rate
}
\description{
Calculate numerical reaction rates
}
\examples{
library(nuclear)

 NumRateHe3He3(1,1)

}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC, and HK
}
\keyword{Nuclear}
\keyword{rate}
