% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfactorTdn.R
\name{sfactorTdn}
\alias{sfactorTdn}
\title{Estimate Astrophysical S-factor}
\format{\describe{
\item{x}{
The function has five arguments:  ecm, e0, gi, gf, ri, rf}
}}
\usage{
sfactorTdn(e0 = e0, gi = gi, gf = gf, ri = ri, rf = rf)
}
\arguments{
\item{e0}{e0}

\item{gi}{gi}

\item{gf}{gf}

\item{r_i}{r_i}

\item{r_f}{r_f}
}
\value{
S-factor
}
\description{
Provides a confusion matrix of classification statistics following logistic regression.
}
\examples{
library(nuclear)

N <- 300
obsx1 <- exp(seq(log(1e-3), log(1),length.out=N))
plot(obsx1,sfactorTdn(obsx1,0.0912,2.93,0.0794,5,5),ylim=c(0,32.5),
     xlim=c(1e-3,1),col="red",cex=1.25,type="l",ylab="S-factor",xlab="E",log="x")

}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{S-factor}
