% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfactorTdn_5p.R
\name{sfactorTdn_5p}
\alias{sfactorTdn_5p}
\alias{sfactorTdn}
\title{Estimate Astrophysical S-factor}
\format{\describe{
\item{x}{
The function has four arguments: ECM, ER, gi, gf, r_i, r_f}
}}
\usage{
sfactorTdn_5p(ECM = ECM, ER = ER, gi = gi, gf = gf, r_i = r_i, r_f = r_f)
}
\arguments{
\item{ECM}{ECM}

\item{ER}{ER}

\item{gi}{gi}

\item{gf}{gf}

\item{r_i}{r_i}

\item{r_f}{r_f}
}
\value{
S-factor
}
\description{
Provides a confusion matrix of classification statistics following logistic regression.
}
\examples{
library(nuclear)

N <- 300
obsx1 <- exp(seq(log(1e-3), log(1),length.out=N))
plot(obsx1,sfactorTdn(obsx1,0.0912,2.93,0.0794),ylim=c(0,32.5),
     xlim=c(1e-3,1),col="red",cex=1.25,type="l",ylab="S-factor",xlab="E",log="x")

}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{S-factor}
