% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma7Benp.R
\name{sigma7Benp}
\alias{sigma7Benp}
\title{Estimate Astrophysical S-factor}
\format{\describe{
\item{x}{
The function has 9  arguments: ecm, e0, ga, gb, ra, rb, jr,la,lb}
}}
\usage{
sigma7Benp(ecm = ecm, e0 = e0, ga = ga, gb = gb,
ra = ra, rb = rb, jr = jr,la = la, lb = lb)
}
\arguments{
\item{ecm}{ecm}

\item{e0}{e0}

\item{ga}{ga}

\item{gb}{gb}

\item{ra}{ra}

\item{rb}{rb}

\item{jr}{jr}

\item{la}{la}

\item{lb}{lb}
}
\value{
Sigma
}
\description{
Provides a sqrt(E)*sigma for Be7(n,p).
}
\examples{
library(nuclear)

N <- 300
obsx1 <- exp(seq(log(1e-5), log(5),length.out=N))
plot(obsx1,sqrt(obsx1)*sigma7Benp(obsx1,0.021,1.6,0.4,5,5,2,0,0),
col="red",cex=1.25,type="l",ylab="Sigma",xlab="E",log="x")

}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{*}
\keyword{SIGMA}
\keyword{sqrt(E)}
