% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma7Benp7mod.R
\name{sigma7Benp7mod}
\alias{sigma7Benp7mod}
\title{Estimate Astrophysical S-factor}
\format{\describe{
\item{x}{
The function has 57 arguments: ecm, e0_1, ga_1, gb_1, ra_1, rb_1, jr_1, la_1, lb_1,
                                    e0_2, ga_2, gb_2, ra_2, rb_2, jr_2, la_2, lb_2,
                                    e0_3, ga_3, gb_3, ra_3, rb_3, jr_3, la_3, lb_3,
                                    e0_4, ga_4, gb_4, ra_4, rb_4, jr_4, la_4, lb_4,
                                    e0_5, ga_5, gb_5, ra_5, rb_5, jr_5, la_5, lb_5,
                                    e0_6, ga_6, gb_6, ra_6, rb_6, jr_6, la_6, lb_6,
                                    e0_7, ga_7, gb_7, ra_7, rb_7, jr_7, la_7, lb_7}
}}
\usage{
sigma7Benp7mod(ecm = ecm, e_01 = e_01, ga_1 = ga_1, gb_1 = gb_1,
ra_1 = ra_1, rb_1 = rb_1, jr_1 = jr_1,la_1 = la_1, lb_1 = lb_1)
}
\arguments{
\item{ecm}{ecm}

\item{e0_1}{e0_1}

\item{ga_1}{ga_1}

\item{g_b}{g_b}

\item{ra_1}{ra_1}

\item{rb_1}{rb_1}

\item{jr_1}{jr_1}

\item{la_1}{la_1}

\item{lb_1}{lb_1}
}
\value{
Sigma
}
\description{
Provides a sqrt(E)*sigma for Be7(n,p).
}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{*}
\keyword{SIGMA}
\keyword{sqrt(E)}
