% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hazan.R
\name{sparse_approx_sdp}
\alias{sparse_approx_sdp}
\title{Hazan's SparseApproxSDP}
\usage{
sparse_approx_sdp(f, gradf, iter_mult, max_its)
}
\arguments{
\item{f}{The function to be minimized, maps S^(n*n) to R}

\item{gradf}{The gradient of the function, maps S^(n*n) to R^(n*n)}

\item{iter_mult}{rARPACK will use iter_mult * k max iterations}

\item{max_its}{Maximum number of iterations}
}
\description{
Uses rARPACK to minimize convex f with given gradient and multiplier (proportional to C_f)
}

