% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfactor3Hedp.R
\name{sfactor3Hedp}
\alias{sfactor3Hedp}
\title{Estimate Astrophysical S-factor}
\format{\describe{
\item{x}{
The function has 6  arguments: ECM, E0, gi, gf, ri, rf, ue}
}}
\usage{
sfactor3Hedp(ecm = ecm, e0 = e0, gi = gi, gf = gf, ri = ri, rf = rf, ue = ue)
}
\arguments{
\item{ecm}{ECM}

\item{e0}{E0}

\item{gi}{gi}

\item{gf}{gf}

\item{ri}{ri}

\item{rf}{rf}

\item{ue}{ue}
}
\value{
S-factor
}
\description{
Provides a confusion matrix of classification statistics following logistic regression.
}
\examples{
library(nuclear)

N <- 300
obsx1 <- exp(seq(log(1e-3), log(1),length.out=N))
plot(obsx1,sfactor3Hedp(obsx1,0.35,1.0085,0.025425),
col="red",cex=1.25,type="l",ylab="S-factor",xlab="E",log="x")

}
\author{
Rafael de Souza, UNC,  and Christian Illiadis, UNC
}
\keyword{S-factor}
